+++ lib/internal/modules/run_main.js
@@ -2,12 +2,12 @@
 
 const {
   ObjectCreate,
+  StringPrototypeEndsWith,
 } = primordials;
 const CJSLoader = require('internal/modules/cjs/loader');
-const { Module, toRealPath } = CJSLoader;
+const { Module, toRealPath, readPackageScope } = CJSLoader;
 const { getOptionValue } = require('internal/options');
 const path = require('path');
-const { shouldUseESMLoader } = require('internal/modules/utils');
 const {
   handleProcessExit,
 } = require('internal/modules/esm/handle_process_exit');
@@ -25,6 +25,27 @@ function resolveMainPath(main) {
     mainPath = toRealPath(mainPath);
 
   return mainPath;
+}
+
+function shouldUseESMLoader(mainPath) {
+  /**
+   * @type {string[]} userLoaders A list of custom loaders registered by the user
+   * (or an empty list when none have been registered).
+   */
+  const userLoaders = getOptionValue('--experimental-loader');
+  if (userLoaders.length > 0)
+    return true;
+  const esModuleSpecifierResolution =
+    getOptionValue('--experimental-specifier-resolution');
+  if (esModuleSpecifierResolution === 'node')
+    return true;
+  // Determine the module format of the main
+  if (mainPath && StringPrototypeEndsWith(mainPath, '.mjs'))
+    return true;
+  if (!mainPath || StringPrototypeEndsWith(mainPath, '.cjs'))
+    return false;
+  const pkg = readPackageScope(mainPath);
+  return pkg && pkg.data.type === 'module';
 }
 
 function runMainESM(mainPath) {
