+++ lib/internal/test_runner/utils.js
@@ -1,6 +1,7 @@
 'use strict';
 const {
   ArrayPrototypePush,
+  ObjectCreate,
   ObjectGetOwnPropertyDescriptor,
   SafePromiseAllReturnArrayLike,
   RegExp,
@@ -9,9 +10,9 @@ const {
 } = primordials;
 const { basename } = require('path');
 const { createWriteStream } = require('fs');
+const { pathToFileURL } = require('internal/url');
 const { createDeferredPromise } = require('internal/util');
 const { getOptionValue } = require('internal/options');
-const { requireOrImport } = require('internal/modules/utils');
 
 const {
   codes: {
@@ -103,7 +104,17 @@ const kDefaultDestination = 'stdout';
 async function getReportersMap(reporters, destinations) {
   return SafePromiseAllReturnArrayLike(reporters, async (name, i) => {
     const destination = kBuiltinDestinations.get(destinations[i]) ?? createWriteStream(destinations[i]);
-    let reporter = await requireOrImport(kBuiltinReporters.get(name) ?? name);
+
+    // Load the test reporter passed to --test-reporter
+    const reporterSpecifier = kBuiltinReporters.get(name) ?? name;
+    let parentURL;
+    try {
+      parentURL = pathToFileURL(process.cwd() + '/').href;
+    } catch {
+      parentURL = 'file:///';
+    }
+    const { esmLoader } = require('internal/process/esm_loader');
+    let reporter = await esmLoader.import(reporterSpecifier, parentURL, ObjectCreate(null));
 
     if (reporter?.default) {
       reporter = reporter.default;
